+++ src/api/commands/user.go
@@ -3,29 +3,47 @@ package commands
 import (
 	"flag"
 	"log"
+	"net/http"
 )
 
 func NewUserCommand() *UserCommand {
-	uc := &UserCommand{
-		fs: flag.NewFlagSet("user", flag.ExitOnError),
+	userCmd := &UserCommand{
+		flagset: flag.NewFlagSet("user", flag.ExitOnError),
 	}
 
-	return uc
+	return userCmd
 }
 
 type UserCommand struct {
-	fs *flag.FlagSet
+	flagset *flag.FlagSet
 }
 
 func (u *UserCommand) Name() string {
-	return u.fs.Name()
+	return u.flagset.Name()
 }
 
 func (u *UserCommand) Init(args []string) error {
-	return u.fs.Parse(args)
+	return u.flagset.Parse(args)
 }
 
 func (u *UserCommand) Run() error {
 	log.Println("user")
+
+	// make a http request to the api
+
+	req, err := http.NewRequest("GET", "http://localhost:8080/api/users", nil)
+
+	if err != nil {
+		return err
+	}
+
+	resp, err := http.DefaultClient.Do(req)
+
+	if err != nil {
+		return err
+	}
+
+	log.Println(resp)
+
 	return nil
 }